+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  - Add msg.event for button clicks.
  - Make Gridstack a normal dep so they can count installs. Bump to v0.5.5
  - Let mousewheel change slider. Issue #575.
+ - Add UTC option to Chart node X-Axis
 
  **Fixes**